Update README.md
Precision agriculture is in trend nowadays. It helps the farmers to get informed decision 
about the farming strategy. Here, we present you a dataset which would allow the users 
to build a predictive model to recommend the most suitable crops to grow in a particular 
farm based on various parameters.